feat: add migrate on deploy and bison config to package.json
This ensures that builds are migrated properly for branches we care about (the staging and prod branches)

In order to properly do this, we add a bison config to package.json. That stores:
* The version used to init the project
* the staging branch
* the production branch

Once we move to Lerna and have a separate Bison package, we wont need the version. But until that time it will be really helpful in helping users ugrade should they need to.